move the fullTypeName onj to external static file
