+++ client/src/components/common/FilterResult/index.js
@@ -2,30 +2,20 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import './style.css';
+import fullTypeName from './staticData';
 
-const fullTypeName = {
-  CBT: 'Cognitive Behavioural Therapy',
-  PD: 'Psychodynamic Therapy',
-  Hu: 'Humanistic psychotherapy',
-  In: 'ntegrative psychotherapy',
-  C: 'Counselling',
-  STPD: 'Short Term Psychodynamic Psychotherapy',
-};
-const FilterResult = props => {
-  const { resultPoints } = props;
-  return (
-    <div className="Results__TherapyType">
-      <h3 className="Results__TherapyType__title">Therapy Type</h3>
-      <ul className="result__list">
-        {Object.keys(resultPoints).map(type => (
-          <li
-            key={type}
-          >{`${fullTypeName[type]}( ${type} ) : ${resultPoints[type]}`}</li>
-        ))}
-      </ul>
-    </div>
-  );
-};
+const FilterResult = ({ resultPoints }) => (
+  <div className="Results__TherapyType">
+    <h3 className="Results__TherapyType__title">Therapy Type</h3>
+    <ul className="result__list">
+      {Object.keys(resultPoints).map(type => (
+        <li
+          key={type}
+        >{`${fullTypeName[type]} ( ${type} ) : ${resultPoints[type]}`}</li>
+      ))}
+    </ul>
+  </div>
+);
 
 FilterResult.propTypes = {
   resultPoints: PropTypes.shape({
