+++ client/src/components/common/FilterResult/staticData.js
@@ -1,0 +1,10 @@
+const fullTypeName = {
+  CBT: 'Cognitive Behavioural Therapy',
+  PD: 'Psychodynamic Therapy',
+  Hu: 'Humanistic psychotherapy',
+  In: 'ntegrative psychotherapy',
+  C: 'Counselling',
+  STPD: 'Short Term Psychodynamic Psychotherapy',
+};
+
+export default fullTypeName;
